requirements.txt
pylint
pycodestyle
pytest
pytest-cov
pytest-env
pytest-pythonpath
pytest-aiohttp
autopep8
sphinx
setuptools
wheel
twine
